# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sayohatchilar o'rtasida tajriba almashish, sayohat rejalarini tuzish va boshqalar bilan bo'lishish imkoniyatini yaratish orqali global sayohat hamjamiyatini shakllantirish.",
        "title": "TravelTales App",
        "contact": {
            "name": "Saidakbar",
            "url": "http://www.support_me_with_smile",
            "email": "\"pardaboyevsaidakbar103@gmail.com\""
        },
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/travel",
    "paths": {
        "/stories": {
            "get": {
                "description": "this is for getting stories information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "summary": "Get stories info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit is required",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page is required",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns stories information",
                        "schema": {
                            "$ref": "#/definitions/stories.ResponseGetStories"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stories/{id}/edit": {
            "put": {
                "description": "this is for editing story",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "summary": "Edit Story",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id is required",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "all params are required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stories.RequestEditStory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns story information",
                        "schema": {
                            "$ref": "#/definitions/stories.ResponseEditStory"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/stories/create": {
            "post": {
                "description": "this is for creating story",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "summary": "Create Story",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id is required",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "all params are required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stories.RequestCreateStory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns story information",
                        "schema": {
                            "$ref": "#/definitions/stories.ResponseCreateStory"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "stories.Author": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "stories.RequestCreateStory": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "location": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "stories.RequestEditStory": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "location": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "stories.ResponseCreateStory": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "stories.ResponseEditStory": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "stories.ResponseGetStories": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "stories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/stories.StoryForGet"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "stories.StoryForGet": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/stories.Author"
                },
                "comments_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "likes_count": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}